backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

spacy==3.7.2
transformers==4.34.1
torch

easyocr
opencv-python-headless
Pillow
exifread

numpy<2.0

moviepy==1.0.3
pydub

en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

openai-whisper